+++ crates/tinywasm/src/func.rs
@@ -2,7 +2,7 @@ use crate::interpreter::stack::{CallFrame, Stack};
 use crate::{log, unlikely, Function};
 use crate::{Error, FuncContext, Result, Store};
 use alloc::{boxed::Box, format, string::String, string::ToString, vec, vec::Vec};
-use tinywasm_types::{FuncType, ModuleInstanceAddr, ValType, WasmValue, WasmFuncRef, WasmExternRef};
+use tinywasm_types::{ExternRef, FuncRef, FuncType, ModuleInstanceAddr, ValType, WasmValue};
 
 #[derive(Debug)]
 /// A function handle
@@ -51,9 +51,9 @@ impl FuncHandle {
         let func_inst = store.get_func(self.addr);
         let wasm_func = match &func_inst.func {
             Function::Host(host_func) => {
-                let func = &host_func.clone().func;
+                let host_func = host_func.clone();
                 let ctx = FuncContext { store, module_addr: self.module_addr };
-                return (func)(ctx, params);
+                return host_func.call(ctx, params);
             }
             Function::Wasm(wasm_func) => wasm_func,
         };
@@ -219,15 +219,15 @@ impl ToValType for f64 {
     }
 }
 
-impl ToValType for WasmExternRef {
+impl ToValType for FuncRef {
     fn to_val_type() -> ValType {
-        ValType::RefExtern
+        ValType::RefFunc
     }
 }
 
-impl ToValType for WasmFuncRef {
+impl ToValType for ExternRef {
     fn to_val_type() -> ValType {
-        ValType::RefFunc
+        ValType::RefExtern
     }
 }
 
@@ -263,15 +263,15 @@ impl_from_wasm_value_tuple_single!(i32);
 impl_from_wasm_value_tuple_single!(i64);
 impl_from_wasm_value_tuple_single!(f32);
 impl_from_wasm_value_tuple_single!(f64);
-impl_from_wasm_value_tuple_single!(WasmFuncRef);
-impl_from_wasm_value_tuple_single!(WasmExternRef);
+impl_from_wasm_value_tuple_single!(FuncRef);
+impl_from_wasm_value_tuple_single!(ExternRef);
 
 impl_into_wasm_value_tuple_single!(i32);
 impl_into_wasm_value_tuple_single!(i64);
 impl_into_wasm_value_tuple_single!(f32);
 impl_into_wasm_value_tuple_single!(f64);
-impl_into_wasm_value_tuple_single!(WasmFuncRef);
-impl_into_wasm_value_tuple_single!(WasmExternRef);
+impl_into_wasm_value_tuple_single!(FuncRef);
+impl_into_wasm_value_tuple_single!(ExternRef);
 
 impl_val_types_from_tuple!(T1);
 impl_val_types_from_tuple!(T1, T2);